Relocate metrics tracker from plugin reporting to shared utilities
