+++ src/plugin/src/reporting/index.js
@@ -1,1 +1,1 @@
-export * from "./metrics.js";
+export { createMetricsTracker } from "../../../shared/metrics-utils.js";
