+++ src/shared/utils/index.js
@@ -7,6 +7,7 @@ export * from "./error.js";
 export * from "./fs.js";
 export * from "./identifier-metadata.js";
 export * from "./json.js";
+export * from "./metrics.js";
 export * from "./line-breaks.js";
 export * from "./number.js";
 export * from "./numeric-options.js";
